Feel its in a good place now.
Still feel its a mit hacky though. With all the numbers for pushing and pulling.
I don't really have a choice right now, other than laying a grid over the whole page, which is not idea.
I'd need to find another way to do this that's more reliable and zooms the way one'd expect.
Also need to add callbacks and author/time tags. But need to genericize them, not tie it to fashun.
Maybe fashun-fork?